+++ client/src/App.tsx
@@ -7,6 +7,7 @@ import {FocusStyleManager, Spinner} from "@blueprintjs/core";
 import SideBar from "./Components/SideBar";
 import CodeViewer from "./Components/CodeViewer";
 import {PromiseWithCancel} from "./Interfaces/PromiseWithCancel";
+import MainPage from "./Components/MainPage";
 
 FocusStyleManager.onlyShowFocusOnTabs();
 
@@ -21,40 +22,6 @@ export default function () {
         pendant: [],
         tops: []
     });
-    const server = process.env.REACT_APP_SERVER_URL;
-    const [coverDepth, setCoverDepth] = useState<number>(1);
-    const [query, setQuery] = useState<PromiseWithCancel<any> | undefined>();
-
-
-    function doFetch(path: string, method: string, body: any, resolve?: (res: any) => void, name?: string) {
-        if (!query) {
-            setKernel({isolated: [], pendant: [], tops: []});
-            setCover({depth: coverDepth, vertices: []});
-            const controller = new AbortController();
-            const signal = controller.signal;
-            const promise = new Promise(async () => {
-                try {
-                    const response = await fetch(server + path, {
-                        method: method,
-                        body: JSON.stringify(body),
-                        signal
-                    });
-                    const data = await response.json();
-                    setQuery(undefined);
-                    if (resolve)
-                        resolve({data: data, query: (promise as PromiseWithCancel<any>)});
-                } catch (ex: any) {
-                    setQuery(undefined);
-                }
-            });
-            (promise as PromiseWithCancel<any>).cancel = () => controller.abort();
-            (promise as PromiseWithCancel<any>).dateTime = new Date();
-            if (name)
-                (promise as PromiseWithCancel<any>).name = name;
-            setQuery((promise as PromiseWithCancel<any>));
-            return (promise as PromiseWithCancel<any>);
-        }
-    }
 
     useEffect(() => {
         centerNodes();
@@ -96,58 +63,28 @@ export default function () {
         }
     }
 
-    const getGraph = () => {
-        return (
-            <Graph
-                id="graph-id"
-                ref={graphRef}
-                data={convertToD3Graph(graph, cover, kernel)}
-                onClickNode={onClickNode}
-                config={{
-                    staticGraph: false,
-                    height: graphBoundingRef.current != null ? graphBoundingRef.current.offsetHeight : 0,
-                    width: graphBoundingRef.current != null ? graphBoundingRef.current.offsetWidth : 0,
-                    minZoom: 0.5,
-                    maxZoom: 8
-                }}
-            />
-        );
-    }
-
     return (
-        <>
-            <div style={{display: "flex", flexDirection: "row-reverse", flex: "auto", overflow: "hidden"}}>
-                <SideBar
-                    data={graph} setData={setGraph}
-                    cover={cover} setCover={setCover}
-                    kernel={kernel} setKernel={setKernel}
-                    coverDepth={coverDepth} setCoverDepth={setCoverDepth}
-                    doFetch={doFetch}
-                    query={query}
-                />
-                <div
-                    style={{overflow: "hidden", margin: "1em", display: "flex", flex: "auto", flexDirection: "column"}}>
-
-                    <div style={{
-                        position: "absolute",
-                        pointerEvents: "none",
-                        opacity: query ? 1 : 0,
-                        transition: query ? "opacity 0.1s" : "opacity 0.3s",
-                        transitionDelay: query ? "0.05s" : "0.2s"
-                    }}>
-                        <Spinner/>
-                    </div>
-                    <div className="container__graph-area" ref={graphBoundingRef}>
-                        {getGraph()}
-                    </div>
-                    <CodeViewer
-                        data={graph}
-                        setData={setGraph}
-                        graphElement={getGraph()}
-                        graphBoundingRef={graphBoundingRef}
-                        query={query}/>
+            <MainPage
+                graphBoundingRef={graphBoundingRef}
+                graphRef={graphRef}
+                data={graph} setData={setGraph}
+                cover={cover} setCover={setCover}
+                kernel={kernel} setKernel={setKernel}>
+                <div className="container__graph-area" ref={graphBoundingRef}>
+                    <Graph
+                        id="graph-id"
+                        ref={graphRef}
+                        data={convertToD3Graph(graph, cover, kernel)}
+                        onClickNode={onClickNode}
+                        config={{
+                            staticGraph: false,
+                            height: graphBoundingRef.current != null ? graphBoundingRef.current.offsetHeight : 0,
+                            width: graphBoundingRef.current != null ? graphBoundingRef.current.offsetWidth : 0,
+                            minZoom: 0.5,
+                            maxZoom: 8
+                        }}
+                    />
                 </div>
-            </div>
-        </>
+            </MainPage>
     );
 }
