+++ client/src/index.tsx
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
-import NavBar from "./NavBar/NavBar";
+import NavBar from "./Components/NavBar";
 
 
 ReactDOM.render(
